fixing refresh logic
